Un comment line 50

+++ cogs/events/auto_mod.py
@@ -48,23 +48,22 @@ class AutoModEvent(commands.Cog):
         if status == 'enabled':
             member = msg.guild.get_member(msg.author.id)
             if member is not None:
-                # if not member.guild_permissions.manage_messages: #If message author doesn't have manage_messages permission
-                                       
-                if str(msg.guild.id) in cache:
-                    blacklist = cache[str(msg.guild.id)]
+                if not member.guild_permissions.manage_messages: #If message author doesn't have manage_messages permission  
+                    if str(msg.guild.id) in cache:
+                        blacklist = cache[str(msg.guild.id)]
 
-                else:
-                    blacklist = await self.get_blacklist(msg.guild.id)
-                    cache[str(msg.guild.id)] = blacklist
-                    
-                for word in blacklist:
-                    if word.lower() in msg.content.lower():
-                        await msg.delete()
-                        try:
-                            await msg.author.send(f"Hey! That word is not allowed in {msg.guild.name}")
-                        except:
-                            pass
-                        break
+                    else:
+                        blacklist = await self.get_blacklist(msg.guild.id)
+                        cache[str(msg.guild.id)] = blacklist
+                        
+                    for word in blacklist:
+                        if word.lower() in msg.content.lower():
+                            await msg.delete()
+                            try:
+                                await msg.author.send(f"Hey! That word is not allowed in {msg.guild.name}")
+                            except:
+                                pass
+                            break
 
 
 def setup(client):
